tsdb/chunkenc: Create RunLengthEncoding chunk

+++ tsdb/chunkenc/doubledelta.go
@@ -1,3 +1,16 @@
+// Copyright 2017 The Prometheus Authors
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package chunkenc
 
 import (
@@ -90,7 +103,7 @@ func (a *doubleDeltaAppender) Append(t int64, vf float64) {
 		}
 	} else if num == 1 {
 		tDelta := uint64(t - a.t)
-		vDelta := uint64(v - a.v)
+		vDelta := uint64(v) - uint64(a.v)
 
 		buf := make([]byte, binary.MaxVarintLen64)
 		for _, b := range buf[:binary.PutUvarint(buf, tDelta)] {
@@ -125,7 +138,7 @@ func (a *doubleDeltaAppender) Append(t int64, vf float64) {
 		}
 
 		tDelta := uint64(t - a.t)
-		vDelta := uint64(v - a.v)
+		vDelta := uint64(v) - uint64(a.v)
 		tDoD := int64(tDelta - a.tDelta)
 		vDoD := int64(vDelta - a.vDelta)
